fix(ons-navigator): Fix the cancelIfRunning functionality
Currently if a user manages to click twice on a menu element even with  `{ cancelIfRunning : true }`.
The reason for that is that _isPushing is being set in a callback function, meaning that the second execution of the click handler can start before the flag is set to true. This flag isn't used in many places so side effects from this should not exist. Both `core-test` and `e2e-test` show no errors after this change.
The original assignment of _isPushing to true is left unchanged due to the function _pushPageDOM being called in bringPageTop.
Replication of the bug is hard to achieve by hand so it can be done with [this codepen](http://codepen.io/IliaSky/pen/jWGWeg?editors=001) and [this repeater script](https://github.com/OnsenUI/OnsenUI/files/91770/click.record.txt) (.record). The codepen also shows the relative part of the fix, so it can be enabled and seen easily.