feat(ISSUE_TEMPLATE.md): Add initial issue template.
We should at least have something even if it's not perfect. I feel like what our users need is guidelines more than a template. However most users woudn't click the 'please read the guidelines' and even if they do they will find the guidelines for pull requests rather than bug reports. I sort of cheated the system by using an empty link as a comment, but that's all I could think of. Adding more content in the template is not nice imo as it's just annoying to delete.