fix(gesture-detector): Fix defaults so that they don't disable text selection in inputs.
With the current state we are disabling text selection trough `onselectstart` which disables even the children when it's set on the parent with our current version. Since we're already disabling it with `user-select: none` we shouldn't rely on this feature from the gesture detector. And since we currently have a gesture detector in the body that makes selection impossible in all inputs and textareas. (Actually you can still select it, but it disables the `select all` option, which people probably may want to use in their phones.) A user noticed this bug in Onsen 1, but there it's easier for them to disable it while in Onsen 2 it's a bit harder. Anyway we shouldn't do this disabling in the first place. I'm leaving it in a comment in case someone decides to change it back later at some point.