fix(ons-alert-dialog): Stop compiling twice.
Currently the code tries to compile twice. @argelius - I saw that you made lots of changing adding and removing one of the compiles. I've added the _compiled attribute once again and forced it to compile only once. Recompiling was causing some issues because the second compile was being called after the dialog is shown, hiding it.