chore(angularjs): Move AngularJS bindings to own package
As per discussion in #2476, this commit moves the AngularJS bindings into a new package
called angularjs-onsenui. For now, the bindings continue to be packaged with the core,
but with an added warning to users that they should take the opportunity to upgrade to
the new package before 2.11.0, when the bindings will be removed from the core package.